modules/fargate/src/requirements.txt
psycopg2-binary==2.9.3
names==0.3.0
